refactor - update imports to use `searchChatContext`
