+++ src/vs/workbench/contrib/search/browser/search.contribution.ts
@@ -32,7 +32,7 @@ import { CommandsRegistry } from '../../../../platform/commands/common/commands.
 import { assertType } from '../../../../base/common/types.js';
 import { getWorkspaceSymbols, IWorkspaceSymbol } from '../common/search.js';
 import * as Constants from '../common/constants.js';
-import { SearchChatContextContribution } from './chatContributions.js';
+import { SearchChatContextContribution } from './searchChatContext.js';
 
 import './searchActionsCopy.js';
 import './searchActionsFind.js';
