+++ src/vs/workbench/contrib/search/browser/searchChatContext.ts
